+++ extensions/spotify-player/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Spotify Player Changelog
 
+## [Feature] - 2022-08-11
+
+- Add new command "My Playlists"
+
 ## [Bug Fixes] - 2022-08-01
 
 - Properly handled states for unauthorized state in menu bar and no-view commands